requirements.txt
logzero==1.5.0
pika==0.12.0
pyaml==17.12.1
PyYAML==3.13
redis==2.10.6
